requirements.txt
absl-py==0.11.0
aiocontextvars==0.2.2
aiohttp==3.7.4
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
einops==0.3.0
entrypoints==0.3
fancycompleter==0.9.1
fsspec==0.8.7
future==0.18.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
greenlet==1.0.0
grpcio==1.35.0
h5py==3.1.0
Hydra==2.5
hydra-core==1.1.0.dev4
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.15
importlib-metadata==3.7.0
importlib-resources==5.1.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==0.5.4
multidict==5.1.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
neovim==0.2.1
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.1.0.dev26
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
ordered-set==4.0.2
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrepl==0.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.1.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.1.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
submitit==1.1.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.4.5
torch==1.7.0
torchmeta==1.7.0
torchmetrics==0.2.0
torchvision==0.8.1
tornado==6.1
tqdm==4.58.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wmctrl==0.3
yapf==0.30.0
yarl==1.6.3
zipp==3.4.0
